requirements.txt
django
djangorestframework
drf-yasg
djangorestframework_simplejwt
psycopg2 # For PostgresSQL